Relocate recommendation tests under subdirectory
