+++ caseworker/f680/recommendation/tests/test_views.py
@@ -56,7 +56,7 @@ def recommendation(current_user, admin_team):
         {
             "created_at": "2021-10-16T23:48:39.486679+01:00",
             "denial_reasons": [],
-            "id": "429c5596-fe8b-4540-988b-c37805cd08de",
+            "id": "429c5596-fe8b-4540-988b-c37805cd08de",  # /PS-IGNORE
             "level": "user",
             "note": "additional notes",
             "text": "No concerns",
